--- src/main/java/com/karankumar/booksapi/model/format/FormatType.java
@@ -1,36 +1,0 @@
-/*
- * Copyright (C) 2021  Karan Kumar
- *
- * This program is free software: you can redistribute it and/or modify it under the terms of the
- * GNU General Public License as published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
-
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- * PURPOSE.  See the GNU General Public License for more details.
-
- * You should have received a copy of the GNU General Public License along with this program.
- * If not, see <https://www.gnu.org/licenses/>.
- */
-
-package com.karankumar.booksapi.model.format;
-
-public enum FormatType {
-    // IDs correspond the IDs in the associated database table
-    EBOOK("eBook", 1),
-    HARDCOVER("Hardcover", 2),
-    PAPERBACK("Paperback", 3);
-
-    private final String type;
-    private final int id;
-
-    FormatType(String type, int id) {
-        this.type = type;
-        this.id = id;
-    }
-
-    @Override
-    public String toString() {
-        return type;
-    }
-}
