+++ src/main/resources/db/migration/V012__CREATE_COVER.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS book_cover (
+CREATE TABLE IF NOT EXISTS cover (
     id               SERIAL  NOT NULL PRIMARY KEY,
     small_file_type  INTEGER REFERENCES cover_file_type(id),
     medium_file_type INTEGER REFERENCES cover_file_type(id),
