--- src/main/resources/db/migration/V013__ALTER_BOOK_ADD_BOOK_COVER.sql
@@ -1,3 +1,0 @@
-ALTER TABLE book
-ADD COLUMN book_cover_id INTEGER REFERENCES book_cover(id);
-
