+++ src/main/resources/db/migration/V013__ALTER_BOOK_ADD_COVER.sql
@@ -1,0 +1,3 @@
+ALTER TABLE book
+ADD COLUMN cover_id INTEGER REFERENCES cover(id);
+
