--- src/main/resources/db/migration/V17__CREATE_FORMAT.sql
@@ -1,8 +1,0 @@
-CREATE TABLE format (
-    id   SERIAL      NOT NULL PRIMARY KEY,
-    type VARCHAR(20) NOT NULL UNIQUE
-);
-
-INSERT INTO format (type) VALUES ('eBook');
-INSERT INTO format (type) VALUES ('Hardcover');
-INSERT INTO format (type) VALUES ('Paperback');
