--- src/main/resources/schema/bookformat.graphqls
@@ -1,5 +1,0 @@
-enum BookFormat {
-    EBOOK,
-    HARDCOVER,
-    PAPERBACK
-}
